scripts/requirements.txt
torch
coremltools
transformers
fugashi
unidic_lite
Pillow